requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.0.1